ISSUE #231

+++ src/Soap/SoapBase.php
@@ -572,26 +572,34 @@ abstract class SoapBase implements SoapInterface
      */
     public function removeTemporarilyFiles()
     {
-        if (empty($this->filesystem) || empty($this->certsdir)) {
-            return;
-        }
-        //remove os certificados
-        $this->filesystem->delete($this->certfile);
-        $this->filesystem->delete($this->prifile);
-        $this->filesystem->delete($this->pubfile);
-        //remove todos os arquivos antigos
-        $contents = $this->filesystem->listContents($this->certsdir, true);
-        $dt = new \DateTime();
-        $tint = new \DateInterval("PT".$this->waitingTime."M");
-        $tint->invert = 1;
-        $tsLimit = $dt->add($tint)->getTimestamp();
-        foreach ($contents as $item) {
-            if ($item['type'] == 'file') {
-                $timestamp = $this->filesystem->getTimestamp($item['path']);
-                if ($timestamp < $tsLimit) {
-                    $this->filesystem->delete($item['path']);
+        try {
+            if (empty($this->filesystem) || empty($this->certsdir)) {
+                return;
+            }
+            //remove os certificados
+            $this->filesystem->delete($this->certfile);
+            $this->filesystem->delete($this->prifile);
+            $this->filesystem->delete($this->pubfile);
+            //remove todos os arquivos antigos
+            $contents = $this->filesystem->listContents($this->certsdir, true);
+            $dt = new \DateTime();
+            $tint = new \DateInterval("PT".$this->waitingTime."M");
+            $tint->invert = 1;
+            $tsLimit = $dt->add($tint)->getTimestamp();
+            foreach ($contents as $item) {
+                if ($item['type'] == 'file') {
+                    if ($this->filesystem->fileExists($item['path'])) {
+                        $timestamp = $this->filesystem->getTimestamp($item['path']);
+                        if ($timestamp < $tsLimit) {
+                            $this->filesystem->delete($item['path']);
+                        }
+                    }
                 }
             }
+        } catch (\Throwable $e) {
+            //impedir de ocorrer exception em ambientes muito comporrentes
+            //porem nesses casos devem ser feitas limpezas periodicas caso
+            //não seja usado o diretorio /tmp pois não será auto limpante
         }
     }
